+++ NEWS.md
@@ -2,6 +2,10 @@
 
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
+# topics 0.30.5
+- plots are not added as a list (and not only saved to the folder)
+
 # topics 0.30.4
 - scaling controls with scale instead of manually resulting in sligthtly different estimates. (but still same p-value and t-values)
 - removed ridge regression, t-test and correlation codes since they did not work